can run other files

+++ src/main/kotlin/BuiltinFunctions.kt
@@ -35,8 +35,8 @@ object BuiltinFunctions {
         val arg = args[0]
         return when(arg) {
             is BooleanValue -> arg
-            is IntegerValue -> FloatValue(kotlin.math.sqrt(arg.v as Double))
-            is FloatValue -> FloatValue(kotlin.math.sqrt(arg.v as Double))
+            is IntegerValue -> FloatValue(kotlin.math.sqrt(arg.v.toDouble()))
+            is FloatValue -> FloatValue(kotlin.math.sqrt(arg.v))
             is UnitNumericValue -> UnitNumericValue(sqrt(listOf(arg.n)) as NumericValue, arg.u.power(FloatValue(0.5)))
             else -> TODO("make an error for this")
         }
